refactor: Implement new project layout and structure
