+++ app/domains/common/utils.py
