+++ app/domains/nanomag/router.py
