+++ app/domains/nanozymes/service.py
@@ -1,12 +1,12 @@
 from sqlalchemy.orm import Session
 from typing import List, Dict, Any, Optional
-from app.api.v1.common.utils import build_filtered_query
+from app.domains.common.utils import build_filtered_query
 from sqlalchemy import text
 
-class SeltoxService:
+class NanozymesService:
     @staticmethod
     def get_all_data(db: Session, nanoparticle: Optional[str] = None) -> List[Dict[str, Any]]:
-        table_name = "dbt_serving.serving_all_data_seltox"
+        table_name = "dbt_serving.serving_all_data_nanozymes"
         filters = {"nanoparticle": nanoparticle}
         query, params = build_filtered_query(table_name, filters)
         result = db.execute(query, params)
@@ -14,21 +14,21 @@ class SeltoxService:
 
     @staticmethod
     def get_column_stats(db: Session) -> List[Dict[str, Any]]:
-        table_name = "dbt_serving.serving_analytics_column_stats_seltox"
+        table_name = "dbt_serving.serving_analytics_column_stats_nanozymes"
         query = text(f"SELECT * FROM {table_name}")
         result = db.execute(query)
         return [dict(row._mapping) for row in result]
 
     @staticmethod
     def get_row_stats(db: Session) -> List[Dict[str, Any]]:
-        table_name = "dbt_serving.serving_analytics_row_stats_seltox"
+        table_name = "dbt_serving.serving_analytics_row_stats_nanozymes"
         query = text(f"SELECT * FROM {table_name}")
         result = db.execute(query)
         return [dict(row._mapping) for row in result]
 
     @staticmethod
     def get_top_categories(db: Session) -> List[Dict[str, Any]]:
-        table_name = "dbt_serving.serving_analytics_top_categories_seltox"
+        table_name = "dbt_serving.serving_analytics_top_categories_nanozymes"
         query = text(f"SELECT * FROM {table_name}")
         result = db.execute(query)
         return [dict(row._mapping) for row in result]
