+++ app/domains/synergy/router.py
