+++ main.py
@@ -1,5 +1,5 @@
 from fastapi import FastAPI
-from app.config import settings
+from config import settings
 
 # Создаем экземпляр приложения
 app = FastAPI(
@@ -17,23 +17,23 @@ def include_routers():
     Импортирует и подключает роутеры для каждого домена данных.
     """
     # 1.  Cytotox
-    from app.api.v1.cytotox.router import router as cytotox_router
+    from app.domains.cytotox.router import router as cytotox_router
     app.include_router(cytotox_router, prefix="/api/v1/cytotox", tags=["Cytotox"])
 
     # 2.  Nanomag
-    from app.api.v1.nanomag.router import router as nanomag_router
+    from app.domains.nanomag.router import router as nanomag_router
     app.include_router(nanomag_router, prefix="/api/v1/nanomag", tags=["Nanomag"])
 
     # 3.  Nanozymes
-    from app.api.v1.nanozymes.router import router as nanozymes_router
+    from app.domains.nanozymes.router import router as nanozymes_router
     app.include_router(nanozymes_router, prefix="/api/v1/nanozymes", tags=["Nanozymes"])
 
     # 4.  Seltox
-    from app.api.v1.seltox.router import router as seltox_router
+    from app.domains.seltox.router import router as seltox_router
     app.include_router(seltox_router, prefix="/api/v1/seltox", tags=["Seltox"])
 
     # 5.  Synergy
-    from app.api.v1.synergy.router import router as synergy_router
+    from app.domains.synergy.router import router as synergy_router
     app.include_router(synergy_router, prefix="/api/v1/synergy", tags=["Synergy"])
